+++ lib/poolex.ex
@@ -18,7 +18,7 @@ defmodule Poolex do
   Then you can execute any code on the workers with `run/3`:
 
   ```elixir
-  Poolex.run(:worker_pool, &(is_pid?(&1)), timeout: 1_000)
+  Poolex.run(:worker_pool, &(is_pid?(&1)), checkout_timeout: 1_000)
   {:ok, true}
   ```
 
@@ -27,14 +27,14 @@ defmodule Poolex do
 
   use GenServer, shutdown: :infinity
 
-  alias Poolex.BusyWorkers
-  alias Poolex.DebugInfo
-  alias Poolex.IdleWorkers
-  alias Poolex.Monitoring
-  alias Poolex.State
-  alias Poolex.WaitingCallers
+  alias Poolex.Private.BusyWorkers
+  alias Poolex.Private.DebugInfo
+  alias Poolex.Private.IdleWorkers
+  alias Poolex.Private.Monitoring
+  alias Poolex.Private.State
+  alias Poolex.Private.WaitingCallers
 
-  @default_wait_timeout :timer.seconds(5)
+  @default_checkout_timeout :timer.seconds(5)
   @poolex_options_table """
   | Option                 | Description                                          | Example               | Default value                     |
   |------------------------|------------------------------------------------------|-----------------------|-----------------------------------|
@@ -75,18 +75,11 @@ defmodule Poolex do
   @type worker() :: pid()
 
   @typedoc """
-  Tuple describing the `caller`.
-
-  **Callers** are processes that have requested to get a worker.
+  | Option  | Description                                        | Example  | Default value                           |
+  |---------|----------------------------------------------------|----------|-----------------------------------------|
+  | checkout_timeout | How long we can wait for a worker on the call site | `60_000` | `#{@default_checkout_timeout}` |
   """
-  @type caller() :: GenServer.from()
-
-  @typedoc """
-  | Option  | Description                                        | Example  | Default value              |
-  |---------|----------------------------------------------------|----------|----------------------------|
-  | timeout | How long we can wait for a worker on the call site | `60_000` | `#{@default_wait_timeout}` |
-  """
-  @type run_option() :: {:timeout, timeout()}
+  @type run_option() :: {:checkout_timeout, timeout()}
 
   @doc """
   Starts a Poolex process without links (outside of a supervision tree).
@@ -96,7 +89,7 @@ defmodule Poolex do
   ## Examples
 
       iex> Poolex.start(pool_id: :my_pool, worker_module: Agent, worker_args: [fn -> 0 end], workers_count: 5)
-      iex> %Poolex.State{worker_module: worker_module} = Poolex.get_state(:my_pool)
+      iex> %Poolex.Private.State{worker_module: worker_module} = Poolex.get_state(:my_pool)
       iex> worker_module
       Agent
   """
@@ -120,7 +113,7 @@ defmodule Poolex do
   ## Examples
 
       iex> Poolex.start_link(pool_id: :other_pool, worker_module: Agent, worker_args: [fn -> 0 end], workers_count: 5)
-      iex> %Poolex.State{worker_module: worker_module} = Poolex.get_state(:other_pool)
+      iex> %Poolex.Private.State{worker_module: worker_module} = Poolex.get_state(:other_pool)
       iex> worker_module
       Agent
   """
@@ -154,58 +147,55 @@ defmodule Poolex do
   end
 
   @doc """
-  Same as `run!/3` but handles runtime_errors.
+  The main function for working with the pool.
+
+  It takes a pool identifier, a function that takes a worker process id as an argument and returns any value.
+  When executed, an attempt is made to find a free worker with specified timeout (5 seconds by default).
+  You can set the timeout using the `checkout_timeout` option.
 
   Returns:
-    * `{:runtime_error, reason}` on errors.
-    * `:all_workers_are_busy` if no free worker was found before the timeout.
-
-  See `run!/3` for more information.
+    * `{:ok, result}` if the worker was found and the function was executed successfully.
+    * `{:error, :checkout_timeout}` if no free worker was found before the timeout.
 
   ## Examples
 
       iex> Poolex.start_link(pool_id: :some_pool, worker_module: Agent, worker_args: [fn -> 5 end], workers_count: 1)
-      iex> Poolex.run(:some_pool, fn _pid -> raise RuntimeError end)
-      {:runtime_error, %RuntimeError{message: "runtime error"}}
       iex> Poolex.run(:some_pool, fn pid -> Agent.get(pid, &(&1)) end)
       {:ok, 5}
   """
   @spec run(pool_id(), (worker :: pid() -> any()), list(run_option())) ::
-          {:ok, any()} | :all_workers_are_busy | {:runtime_error, any()}
+          {:ok, any()} | {:error, :checkout_timeout}
   def run(pool_id, fun, options \\ []) do
-    {:ok, run!(pool_id, fun, options)}
-  rescue
-    runtime_error -> {:runtime_error, runtime_error}
-  catch
-    :exit, {:timeout, _meta} -> :all_workers_are_busy
-    :exit, reason -> {:runtime_error, reason}
+    checkout_timeout = Keyword.get(options, :checkout_timeout, @default_checkout_timeout)
+
+    case get_idle_worker(pool_id, checkout_timeout) do
+      {:ok, worker_pid} ->
+        monitor_process = monitor_caller(pool_id, self(), worker_pid)
+
+        try do
+          {:ok, fun.(worker_pid)}
+        after
+          Process.exit(monitor_process, :kill)
+          GenServer.cast(pool_id, {:release_busy_worker, worker_pid})
+        end
+
+      {:error, :checkout_timeout} ->
+        {:error, :checkout_timeout}
+    end
   end
 
-  @doc """
-  The main function for working with the pool.
-
-  When executed, an attempt is made to obtain a worker with the specified timeout (5 seconds by default).
-  In case of successful execution of the passed function, the result will be returned, otherwise an error will be raised.
-
-  ## Examples
-
-      iex> Poolex.start_link(pool_id: :some_pool, worker_module: Agent, worker_args: [fn -> 5 end], workers_count: 1)
-      iex> Poolex.run!(:some_pool, fn pid -> Agent.get(pid, &(&1)) end)
-      5
-  """
-  @spec run!(pool_id(), (worker :: pid() -> any()), list(run_option())) :: any()
-  def run!(pool_id, fun, options \\ []) do
-    timeout = Keyword.get(options, :timeout, @default_wait_timeout)
-
-    {:ok, pid} = GenServer.call(pool_id, :get_idle_worker, timeout)
-
-    monitor_process = monitor_caller(pool_id, self(), pid)
+  @spec get_idle_worker(pool_id(), timeout()) :: {:ok, worker()} | {:error, :checkout_timeout}
+  defp get_idle_worker(pool_id, checkout_timeout) do
+    caller_reference = make_ref()
 
     try do
-      fun.(pid)
+      GenServer.call(pool_id, {:get_idle_worker, caller_reference}, checkout_timeout)
+    catch
+      :exit,
+      {:timeout, {GenServer, :call, [_pool_id, {:get_idle_worker, ^caller_reference}, _timeout]}} ->
+        {:error, :checkout_timeout}
     after
-      Process.exit(monitor_process, :kill)
-      GenServer.cast(pool_id, {:release_busy_worker, pid})
+      GenServer.cast(pool_id, {:cancel_waiting, caller_reference})
     end
   end
 
@@ -217,7 +207,7 @@ defmodule Poolex do
   ## Examples
 
       iex> Poolex.start(pool_id: :my_pool_2, worker_module: Agent, worker_args: [fn -> 0 end], workers_count: 5)
-      iex> state = %Poolex.State{} = Poolex.get_state(:my_pool_2)
+      iex> state = %Poolex.Private.State{} = Poolex.get_state(:my_pool_2)
       iex> state.worker_module
       Agent
       iex> is_pid(state.supervisor)
@@ -249,7 +239,7 @@ defmodule Poolex do
   ## Examples
 
       iex> Poolex.start(pool_id: :my_pool_3, worker_module: Agent, worker_args: [fn -> 0 end], workers_count: 5)
-      iex> debug_info = %Poolex.DebugInfo{} = Poolex.get_debug_info(:my_pool_3)
+      iex> debug_info = %Poolex.Private.DebugInfo{} = Poolex.get_debug_info(:my_pool_3)
       iex> debug_info.busy_workers_count
       0
       iex> debug_info.idle_workers_count
@@ -279,25 +269,28 @@ defmodule Poolex do
       Keyword.get(opts, :waiting_callers_impl, Poolex.Callers.Impl.ErlangQueue)
 
     {:ok, monitor_id} = Monitoring.init(pool_id)
-    {:ok, supervisor} = Poolex.Supervisor.start_link()
+    {:ok, supervisor} = Poolex.Private.Supervisor.start_link()
 
-    state = %State{
-      busy_workers_impl: busy_workers_impl,
-      busy_workers_state: BusyWorkers.init(busy_workers_impl),
-      idle_workers_impl: idle_workers_impl,
-      max_overflow: max_overflow,
-      monitor_id: monitor_id,
-      pool_id: pool_id,
-      supervisor: supervisor,
-      waiting_callers_impl: waiting_callers_impl,
-      waiting_callers_state: WaitingCallers.init(waiting_callers_impl),
-      worker_args: worker_args,
-      worker_module: worker_module,
-      worker_start_fun: worker_start_fun
-    }
+    state =
+      %State{
+        max_overflow: max_overflow,
+        monitor_id: monitor_id,
+        pool_id: pool_id,
+        supervisor: supervisor,
+        worker_args: worker_args,
+        worker_module: worker_module,
+        worker_start_fun: worker_start_fun
+      }
 
-    worker_pids = start_workers(workers_count, state, monitor_id)
-    {:ok, %State{state | idle_workers_state: IdleWorkers.init(idle_workers_impl, worker_pids)}}
+    initial_workers_pids = start_workers(workers_count, state, monitor_id)
+
+    state =
+      state
+      |> IdleWorkers.init(idle_workers_impl, initial_workers_pids)
+      |> BusyWorkers.init(busy_workers_impl)
+      |> WaitingCallers.init(waiting_callers_impl)
+
+    {:ok, state}
   end
 
   @spec start_workers(non_neg_integer(), State.t(), Monitoring.monitor_id()) :: [pid]
@@ -334,33 +327,29 @@ defmodule Poolex do
   end
 
   @impl GenServer
-  def handle_call(:get_idle_worker, {from_pid, _} = caller, %State{} = state) do
-    if IdleWorkers.empty?(state.idle_workers_impl, state.idle_workers_state) do
+  def handle_call({:get_idle_worker, caller_reference}, {from_pid, _} = caller, %State{} = state) do
+    if IdleWorkers.empty?(state) do
       if state.overflow < state.max_overflow do
         {:ok, new_worker} = start_worker(state)
 
         Monitoring.add(state.monitor_id, new_worker, :worker)
 
-        state = add_worker_to_busy_workers(state, new_worker)
+        state = BusyWorkers.add(state, new_worker)
 
         {:reply, {:ok, new_worker}, %State{state | overflow: state.overflow + 1}}
       else
         Monitoring.add(state.monitor_id, from_pid, :waiting_caller)
 
-        new_callers_state =
-          WaitingCallers.add(state.waiting_callers_impl, state.waiting_callers_state, caller)
+        state =
+          WaitingCallers.add(state, %Poolex.Caller{reference: caller_reference, from: caller})
 
-        {:noreply, %{state | waiting_callers_state: new_callers_state}}
+        {:noreply, state}
       end
     else
-      {idle_worker_pid, new_idle_workers_state} =
-        IdleWorkers.pop(state.idle_workers_impl, state.idle_workers_state)
+      {idle_worker_pid, state} = IdleWorkers.pop(state)
+      state = BusyWorkers.add(state, idle_worker_pid)
 
-      state = add_worker_to_busy_workers(state, idle_worker_pid)
-
-      new_state = %State{state | idle_workers_state: new_idle_workers_state}
-
-      {:reply, {:ok, idle_worker_pid}, new_state}
+      {:reply, {:ok, idle_worker_pid}, state}
     end
   end
 
@@ -368,18 +357,17 @@ defmodule Poolex do
     {:reply, state, state}
   end
 
-  def handle_call(:get_debug_info, _form, %Poolex.State{} = state) do
+  def handle_call(:get_debug_info, _form, %State{} = state) do
     debug_info = %DebugInfo{
-      busy_workers_count: BusyWorkers.count(state.busy_workers_impl, state.busy_workers_state),
+      busy_workers_count: BusyWorkers.count(state),
       busy_workers_impl: state.busy_workers_impl,
-      busy_workers_pids: BusyWorkers.to_list(state.busy_workers_impl, state.busy_workers_state),
-      idle_workers_count: IdleWorkers.count(state.idle_workers_impl, state.idle_workers_state),
+      busy_workers_pids: BusyWorkers.to_list(state),
+      idle_workers_count: IdleWorkers.count(state),
       idle_workers_impl: state.idle_workers_impl,
-      idle_workers_pids: IdleWorkers.to_list(state.idle_workers_impl, state.idle_workers_state),
+      idle_workers_pids: IdleWorkers.to_list(state),
       max_overflow: state.max_overflow,
       overflow: state.overflow,
-      waiting_callers:
-        WaitingCallers.to_list(state.waiting_callers_impl, state.waiting_callers_state),
+      waiting_callers: WaitingCallers.to_list(state),
       waiting_callers_impl: state.waiting_callers_impl,
       worker_args: state.worker_args,
       worker_module: state.worker_module,
@@ -391,13 +379,18 @@ defmodule Poolex do
 
   @impl GenServer
   def handle_cast({:release_busy_worker, worker}, %State{} = state) do
-    if WaitingCallers.empty?(state.waiting_callers_impl, state.waiting_callers_state) do
+    if WaitingCallers.empty?(state) do
       new_state = release_busy_worker(state, worker)
       {:noreply, new_state}
     else
       new_state = provide_worker_to_waiting_caller(state, worker)
       {:noreply, new_state}
     end
+  end
+
+  @impl GenServer
+  def handle_cast({:cancel_waiting, caller_reference}, %State{} = state) do
+    {:noreply, WaitingCallers.remove_by_reference(state, caller_reference)}
   end
 
   @impl GenServer
@@ -416,15 +409,15 @@ defmodule Poolex do
 
   @spec release_busy_worker(State.t(), worker()) :: State.t()
   defp release_busy_worker(%State{} = state, worker) do
-    if BusyWorkers.member?(state.busy_workers_impl, state.busy_workers_state, worker) do
-      state = remove_worker_from_busy_workers(state, worker)
+    if BusyWorkers.member?(state, worker) do
+      state = BusyWorkers.remove(state, worker)
 
       if state.overflow > 0 do
         stop_worker(state.supervisor, worker)
 
         state
       else
-        add_worker_to_idle_workers(state, worker)
+        IdleWorkers.add(state, worker)
       end
     else
       state
@@ -433,74 +426,21 @@ defmodule Poolex do
 
   @spec provide_worker_to_waiting_caller(State.t(), worker()) :: State.t()
   defp provide_worker_to_waiting_caller(%State{} = state, worker) do
-    {caller, new_waiting_callers_state} =
-      WaitingCallers.pop(state.waiting_callers_impl, state.waiting_callers_state)
+    {caller, state} = WaitingCallers.pop(state)
 
-    GenServer.reply(caller, {:ok, worker})
+    GenServer.reply(caller.from, {:ok, worker})
 
-    %{state | waiting_callers_state: new_waiting_callers_state}
-  end
-
-  @spec add_worker_to_busy_workers(State.t(), worker()) :: State.t()
-  defp add_worker_to_busy_workers(%State{} = state, worker) do
-    %State{
-      state
-      | busy_workers_state:
-          BusyWorkers.add(
-            state.busy_workers_impl,
-            state.busy_workers_state,
-            worker
-          )
-    }
-  end
-
-  @spec add_worker_to_idle_workers(State.t(), worker()) :: State.t()
-  defp add_worker_to_idle_workers(%State{} = state, worker) do
-    %State{
-      state
-      | idle_workers_state:
-          IdleWorkers.add(
-            state.idle_workers_impl,
-            state.idle_workers_state,
-            worker
-          )
-    }
-  end
-
-  @spec remove_worker_from_busy_workers(State.t(), worker()) :: State.t()
-  defp remove_worker_from_busy_workers(%State{} = state, worker) do
-    %State{
-      state
-      | busy_workers_state:
-          BusyWorkers.remove(
-            state.busy_workers_impl,
-            state.busy_workers_state,
-            worker
-          )
-    }
-  end
-
-  @spec remove_worker_from_idle_workers(State.t(), worker()) :: State.t()
-  defp remove_worker_from_idle_workers(%State{} = state, worker) do
-    %State{
-      state
-      | idle_workers_state:
-          IdleWorkers.remove(
-            state.idle_workers_impl,
-            state.idle_workers_state,
-            worker
-          )
-    }
+    state
   end
 
   @spec handle_down_worker(State.t(), pid()) :: State.t()
   defp handle_down_worker(%State{} = state, dead_process_pid) do
     state =
       state
-      |> remove_worker_from_idle_workers(dead_process_pid)
-      |> remove_worker_from_busy_workers(dead_process_pid)
+      |> IdleWorkers.remove(dead_process_pid)
+      |> BusyWorkers.remove(dead_process_pid)
 
-    if WaitingCallers.empty?(state.waiting_callers_impl, state.waiting_callers_state) do
+    if WaitingCallers.empty?(state) do
       if state.overflow > 0 do
         %State{state | overflow: state.overflow - 1}
       else
@@ -508,28 +448,21 @@ defmodule Poolex do
 
         Monitoring.add(state.monitor_id, new_worker, :worker)
 
-        add_worker_to_idle_workers(state, new_worker)
+        IdleWorkers.add(state, new_worker)
       end
     else
       {:ok, new_worker} = start_worker(state)
       Monitoring.add(state.monitor_id, new_worker, :worker)
 
       state
-      |> add_worker_to_busy_workers(new_worker)
+      |> BusyWorkers.add(new_worker)
       |> provide_worker_to_waiting_caller(new_worker)
     end
   end
 
   @spec handle_down_waiting_caller(State.t(), pid()) :: State.t()
   defp handle_down_waiting_caller(%State{} = state, dead_process_pid) do
-    new_waiting_callers_state =
-      WaitingCallers.remove_by_pid(
-        state.waiting_callers_impl,
-        state.waiting_callers_state,
-        dead_process_pid
-      )
-
-    %State{state | waiting_callers_state: new_waiting_callers_state}
+    WaitingCallers.remove_by_pid(state, dead_process_pid)
   end
 
   @impl GenServer