+++ lib/poolex/private/state.ex
@@ -1,35 +1,37 @@
-defmodule Poolex.DebugInfo do
+defmodule Poolex.Private.State do
   @moduledoc """
-  Information with the current state of the pool.
+  Internal structure containing the state of the pool.
 
   Can be used for debugging.
   """
 
-  defstruct busy_workers_count: 0,
-            busy_workers_impl: nil,
-            busy_workers_pids: [],
-            idle_workers_count: 0,
+  defstruct busy_workers_impl: nil,
+            busy_workers_state: nil,
             idle_workers_impl: nil,
-            idle_workers_pids: [],
+            idle_workers_state: nil,
             max_overflow: 0,
+            monitor_id: nil,
             overflow: 0,
+            pool_id: nil,
+            supervisor: nil,
             waiting_callers_impl: nil,
-            waiting_callers: [],
+            waiting_callers_state: nil,
             worker_args: [],
             worker_module: nil,
             worker_start_fun: nil
 
   @type t() :: %__MODULE__{
-          busy_workers_count: non_neg_integer(),
           busy_workers_impl: module(),
-          busy_workers_pids: list(pid()),
-          idle_workers_count: non_neg_integer(),
+          busy_workers_state: Poolex.Workers.Behaviour.state(),
           idle_workers_impl: module(),
-          idle_workers_pids: list(pid()),
+          idle_workers_state: Poolex.Workers.Behaviour.state(),
           max_overflow: non_neg_integer(),
+          monitor_id: atom() | reference(),
           overflow: non_neg_integer(),
+          pool_id: Poolex.pool_id(),
+          supervisor: pid(),
           waiting_callers_impl: module(),
-          waiting_callers: list(pid()),
+          waiting_callers_state: Poolex.Callers.Behaviour.state(),
           worker_args: list(any()),
           worker_module: module(),
           worker_start_fun: atom()